Separate document from buffer and added logger.
